Moving Graphics Thread into specific files
