+++ CMakeLists.txt
@@ -5,10 +5,11 @@ set(EXECUTABLE_OUTPUT_PATH ${PROJECT_BINARY_DIR}/bin)
 
 file(GLOB SRC "src/*.cpp")
 file(GLOB GRAPHICSSRC "src/Graphics/*.cpp")
+file(GLOB LOGICSRC "src/Logic/*.cpp")
 file(GLOB TERRAINSRC "src/Terrain/*.cpp" "src/Terrain/Hlms/*.cpp" "src/Terrain/Hlms/PbsListener/*.cpp")
 file(GLOB COLLISIONSRC "Collision/src/*.cpp")
 file(GLOB COMMONSRC "Common/src/*.cpp" "Common/src/Threading/*.cpp" "Common/src/System/*.cpp")
-add_executable(LeDernierMorkid ${SRC} ${GRAPHICSSRC} ${TERRAINSRC} ${COLLISIONSRC} ${COMMONSRC})
+add_executable(LeDernierMorkid ${SRC} ${GRAPHICSSRC} ${LOGICSRC} ${TERRAINSRC} ${COLLISIONSRC} ${COMMONSRC})
 
 add_custom_target(run COMMAND LeDernierMorkid DEPENDS LeDernierMorkid WORKING_DIRECTORY ${PROJECT_BINARY_DIR}/bin)
